CryptoPkg: Fix memcpy alias prototype to match the aliasee
The memcpy alias in IntrinsicLib declares its count parameter as
unsigned int while the aliasee __memcpy uses size_t. On 64-bit targets
these are different widths. clang 23 diagnoses this through
-Wattribute-alias:

  CopyMem.c:29:17: warning: alias and aliasee have different types
  'void *(void *, const void *, unsigned int)' and
  'void *(void *, const void *, size_t)' [-Wattribute-alias]

and builds that treat warnings as errors fail, e.g. OVMF built with
the CLANGDWARF toolchain in Yocto.

Declare the alias with size_t so both signatures agree.

+++ CryptoPkg/Library/IntrinsicLib/CopyMem.c
@@ -29,9 +29,9 @@ __memcpy (
 __attribute__ ((__alias__ ("__memcpy")))
 void *
 memcpy (
-  void          *dest,
-  const void    *src,
-  unsigned int  count
+  void        *dest,
+  const void  *src,
+  size_t      count
   );
 
 #else
